pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "MOTIFHand"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Hanyang Zhou", email = "zhouhany@usc.edu"}
]
dependencies = [
    "pyzmq",
    "blinker==1.9.0",
    "click==8.1.8",
    "cmapy==0.6.6",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "Flask==3.1.0",
    "fonttools==4.56.0",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "kiwisolver==1.4.8",
    "MarkupSafe==3.0.2",
    "matplotlib==3.10.1",
    "numpy==2.2.3",
    "opencv-python==4.11.0.86",
    "packaging==24.2",
    "pillow==11.1.0",
    "pyparsing==3.2.1",
    "pyserial==3.5",
    "python-dateutil==2.9.0.post0",
    "scipy==1.15.2",
    "six==1.17.0",
    "toml==0.10.2",
    "typing_extensions==4.12.2",
    "Werkzeug==3.1.3",
    "psutil==7.0.0",
    "ahrs>=0.4.0",
    "setuptools>=61.0",
    "mujoco>=3.3.7",
]

[project.optional-dependencies]
raspi = [
    # Raspberry Pi specific dependencies
    "Adafruit-Blinka==8.55.0",
    "adafruit-circuitpython-busdevice==5.2.11",
    "adafruit-circuitpython-connectionmanager==3.1.3",
    "adafruit-circuitpython-mlx90640==1.3.4",
    "adafruit-circuitpython-register==1.10.2",
    "adafruit-circuitpython-requests==4.1.10",
    "adafruit-circuitpython-typing==1.11.2",
    "Adafruit-PlatformDetect==3.77.0",
    "Adafruit-PureIO==1.1.11",
    "binho-host-adapter==0.1.6",
    "picamera==1.13",
    "pithermalcam==1.0.0",
    "pyftdi==0.56.0",
    "pyusb==1.3.1",
    "sysv_ipc==1.1.0",
]
arm = [
    "xarm-python-sdk",
]

[project.urls]
"Homepage" = "https://slurm-lab-usc.github.io/motif-hand/"

[dependency-groups]
dev = [
    "absolufy-imports>=0.3.1",
    "autoflake>=2.3.1",
    "autopep8>=2.3.2",
    "black>=25.9.0",
    "initpy>=0.2.4",
    "isort>=7.0.0",
    "poetry>=2.2.1",
    "pyinit>=0.0.8",
    "pylint>=4.0.1",
    "pyupgrade>=3.21.0",
]


